Rearrange contract parameter structs
