+++ core/src/contracts/pool.rs
@@ -1,4 +1,10 @@
+use std::convert::TryFrom;
+
 use derive_more::From;
+use ergo_lib::ergotree_ir::chain::address::AddressEncoder;
+use ergo_lib::ergotree_ir::chain::address::AddressEncoderError;
+use ergo_lib::ergotree_ir::chain::address::NetworkAddress;
+use ergo_lib::ergotree_ir::chain::address::NetworkPrefix;
 use ergo_lib::ergotree_ir::chain::token::TokenId;
 use ergo_lib::ergotree_ir::ergo_tree::ErgoTree;
 use ergo_lib::ergotree_ir::ergo_tree::ErgoTreeConstantError;
@@ -6,9 +12,9 @@ use ergo_lib::ergotree_ir::mir::constant::TryExtractFromError;
 use ergo_lib::ergotree_ir::mir::constant::TryExtractInto;
 
 use ergo_lib::ergotree_ir::serialization::SigmaParsingError;
+use serde::Deserialize;
+use serde::Serialize;
 use thiserror::Error;
-
-use crate::oracle_config::PoolContractParameters;
 
 #[derive(Clone)]
 pub struct PoolContract {
@@ -116,6 +122,64 @@ impl PoolContract {
             .unwrap()
             .try_extract_into::<TokenId>()
             .unwrap()
+    }
+}
+
+#[derive(Debug, Clone, Deserialize, Serialize)]
+#[serde(
+    try_from = "PoolContractParametersYaml",
+    into = "PoolContractParametersYaml"
+)]
+/// Parameters for the pool contract
+pub struct PoolContractParameters {
+    pub p2s: NetworkAddress,
+    pub refresh_nft_index: usize,
+    pub refresh_nft_token_id: TokenId,
+    pub update_nft_index: usize,
+    pub update_nft_token_id: TokenId,
+}
+
+/// Used to (de)serialize `PoolContractParameters` instance.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+struct PoolContractParametersYaml {
+    p2s: String,
+    on_mainnet: bool,
+    pub refresh_nft_index: usize,
+    pub refresh_nft_token_id: TokenId,
+    pub update_nft_index: usize,
+    pub update_nft_token_id: TokenId,
+}
+
+impl TryFrom<PoolContractParametersYaml> for PoolContractParameters {
+    type Error = AddressEncoderError;
+
+    fn try_from(p: PoolContractParametersYaml) -> Result<Self, Self::Error> {
+        let prefix = if p.on_mainnet {
+            NetworkPrefix::Mainnet
+        } else {
+            NetworkPrefix::Testnet
+        };
+        let address = AddressEncoder::new(prefix).parse_address_from_str(&p.p2s)?;
+        Ok(PoolContractParameters {
+            p2s: NetworkAddress::new(prefix, &address),
+            refresh_nft_index: p.refresh_nft_index,
+            refresh_nft_token_id: p.refresh_nft_token_id,
+            update_nft_index: p.update_nft_index,
+            update_nft_token_id: p.update_nft_token_id,
+        })
+    }
+}
+
+impl From<PoolContractParameters> for PoolContractParametersYaml {
+    fn from(val: PoolContractParameters) -> Self {
+        PoolContractParametersYaml {
+            p2s: val.p2s.to_base58(),
+            on_mainnet: val.p2s.network() == NetworkPrefix::Mainnet,
+            refresh_nft_index: val.refresh_nft_index,
+            refresh_nft_token_id: val.refresh_nft_token_id.clone(),
+            update_nft_index: val.update_nft_index,
+            update_nft_token_id: val.update_nft_token_id.clone(),
+        }
     }
 }
